update view-message route & view
