+++ app/plugins/router.js
@@ -3,7 +3,7 @@ const routes = [].concat(
   require('../routes/healthy'),
   require('../routes/healthz'),
   require('../routes/assets'),
-  require('../routes/message')
+  require('../routes/view-message')
 )
 
 module.exports = {
