+++ app/views/home.njk
@@ -22,7 +22,7 @@
               {% for notification in notifications %}
 
               <tr class="govuk-table__row action"> 
-                <td class="govuk-table__cell"><a class="govuk-link--no-visited-state" href="/messages/message/{{ notification.sbi }}">{{ notification.content.heading }}</a></td>
+                <td class="govuk-table__cell"><a class="govuk-link--no-visited-state" href="/messages/view-message/{{ notification.sbi }}/{{ notification.messageId }}">{{ notification.content.heading }}</a></td>
                 <td class="govuk-table__cell">{{ notification.requestedDate }}</td>
               </tr>
