+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ffc-sfd-messages",
-  "version": "0.1.27",
+  "version": "0.1.28",
   "description": "",
   "homepage": "github.com?owner=defra&repo=ffc-sfd-messages&organization=defra",
   "main": "app/index.js",
